Refactor search events to correct component module
