+++ apps/modernization-ui/src/apps/search/event/components/InvestigationSearch/InvestigationChips.spec.tsx
