+++ apps/modernization-ui/src/apps/search/event/components/InvestigationSearch/InvestigationResults.spec.tsx
