+++ apps/modernization-ui/src/apps/search/event/components/InvestigationSearch/InvestigationResults.tsx
@@ -2,7 +2,7 @@ import { Grid, Pagination } from '@trussworks/react-uswds';
 import { useEffect, useRef } from 'react';
 import { Investigation, PersonParticipation } from 'generated/graphql/schema';
 import { calculateAge } from 'utils/util';
-import '../AdvancedSearch.scss';
+import 'apps/search/advancedSearch/AdvancedSearch.scss';
 import { useNavigate } from 'react-router';
 import { ClassicLink } from 'classic';
 import { NoData } from 'components/NoData';
