+++ apps/modernization-ui/src/apps/search/event/components/labReportSearch/LabReportChips.spec.tsx
